Fix window capture not resetting to previous window when canceling Graphics Capture target selection in Overlay Properties window
...also fix wrong type being used in OutputManager configid_handle IPC handling (unrelated to the issue though)

+++ src/DesktopPlus/OutputManager.cpp
@@ -1646,7 +1646,7 @@ bool OutputManager::HandleIPCMessage(const MSG& msg)
                 ConfigID_Handle handle_id = (ConfigID_Handle)(msg.wParam - configid_bool_MAX - configid_int_MAX - configid_float_MAX);
 
                 uint64_t value = pun_cast<uint64_t, LPARAM>(msg.lParam);
-                intptr_t previous_value = ConfigManager::GetValue(handle_id);
+                uint64_t previous_value = ConfigManager::GetValue(handle_id);
                 ConfigManager::SetValue(handle_id, value);
 
                 switch (handle_id)
